[chore] 파일 구분 처리
